Update Directory.php (#577)
In the delete method of this Directory class, I noticed it uses Laravel Collections, which isn't what Illuminate FileSystemAdapter expects.

If you take a look at the below URL from Illuminate:
https://github.com/laravel/framework/blob/5.6/src/Illuminate/Filesystem/FilesystemAdapter.php#L292

Notice it either takes a string or an array. If you pass in a collection, it'll evaluate "is_array($paths)" to false.
Thus, the delete would always throw a file not found exception due to incorrect handling.