Do not check if Tenancy is installed when running in console
Problem: the `if` first checked if the package is installed, and *then* if we are running in console. That means that the console tries to connect to the database, which may not be the case (e.g. when deploying an application or running in CI).

Solution: the `if` is now inverted and we first check if we are running in console (or tests) **before** checking if the package is installed.